+++ src/bin/todor/main.rs
@@ -6,6 +6,8 @@ mod select;
 mod walk;
 
 use clap::ArgMatches;
+use config::FileFormat;
+use dirs::home_dir;
 use env_logger;
 use failure::Error;
 use ignore::overrides::OverrideBuilder;
@@ -48,7 +50,21 @@ fn main() {
 
 fn run(matches: &ArgMatches) -> Result<i32, Error> {
     let mut builder = TodoRBuilder::new();
-    // TODO: serach for default config file in ~/.config/todor
+
+    // Search for global config file
+    if let Some(mut global_conf) = home_dir() {
+        if cfg!(windows) {
+            global_conf.push(r"AppData\Roaming\lavifb\todor\todor.conf");
+        } else {
+            global_conf.push(".config/todor/todor.conf");
+        }
+
+        info!("searching for global config in '{}'", global_conf.display());
+        if global_conf.exists() && global_conf.metadata().unwrap().len() > 2 {
+            info!("adding global config file...");
+            builder.add_config_file_with_format(global_conf, FileFormat::Hjson)?;
+        }
+    }
 
     if let Some(config_path) = matches.value_of("CONFIG") {
         builder.add_config_file(Path::new(config_path))?;